requirements.txt
tweepy>=4.14.0
anthropic==0.39.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
